backend/requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
google-cloud-bigquery
python-dotenv
requests
